Prevent default export in order to provide typings
