+++ examples/getstarted/src/admin/app.example.js
