--- examples/getstarted/src/admin/app.js
@@ -1,6 +1,0 @@
-export default {
-  config: {
-    locales: ['fr'],
-  },
-  bootstrap() {},
-};
