--- examples/getstarted/src/plugins/myplugin/admin/src/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": false,
-    "module": "es6",
-    "target": "es5",
-    "jsx": "react",
-    "allowJs": true,
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "resolveJsonModule": true,
-    "noEmit": false,
-    "incremental": true
-  },
-  "exclude": ["node_modules", "**/*.test.js", "*.js"]
-}
