+++ examples/kitchensink/src/admin/app.tsx
@@ -1,8 +1,9 @@
-import theme from './extensions/theme'
+import theme from './extensions/theme';
 
-const config =  {
+const config = {
   auth: {
-    logo: 'https://images.unsplash.com/photo-1593642634367-d91a135587b5?ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=750&q=80'
+    logo:
+      'https://images.unsplash.com/photo-1593642634367-d91a135587b5?ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=750&q=80',
   },
   head: {
     favicon:
@@ -26,11 +27,13 @@ const config =  {
   },
   tutorials: false,
   notifications: { release: false },
-}
+};
 
-const bootstrap = (app) => {}
+const bootstrap = (app: any) => {
+  console.log(app);
+};
 
 export default {
   config,
-  bootstrap
+  bootstrap,
 };
